# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Perfil de cliente do e-commerce GANJJ (nome, telefone, endereço).\nCredenciais não ficam aqui: e-mail e senha são do Authorization Service.\n\nPara usar as rotas protegidas, faça login no Authorization\n(POST http://localhost:8081/auth/login), copie o accessToken e informe\nem Authorize no formato: Bearer {token}",
        "title": "GANJJ — Client Service",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/clients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restrito a contas com papel ADMIN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Lista todos os perfis",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Client"
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "403": {
                        "description": "A conta não é ADMIN",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "A conta vem da claim \"sub\" do token, nunca do corpo — assim ninguém cria perfil no lugar de outra pessoa.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "perfil"
                ],
                "summary": "Cria o perfil da conta autenticada",
                "parameters": [
                    {
                        "description": "Dados do perfil",
                        "name": "perfil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.DadosPerfil"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Client"
                        }
                    },
                    "400": {
                        "description": "Campos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "409": {
                        "description": "A conta já possui perfil",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    }
                }
            }
        },
        "/clients/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "perfil"
                ],
                "summary": "Perfil da conta autenticada",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Client"
                        }
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "404": {
                        "description": "A conta ainda não tem perfil",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "perfil"
                ],
                "summary": "Atualiza o próprio perfil",
                "parameters": [
                    {
                        "description": "Dados do perfil",
                        "name": "perfil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.DadosPerfil"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Client"
                        }
                    },
                    "400": {
                        "description": "Campos inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "404": {
                        "description": "A conta ainda não tem perfil",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "A conta em si continua existindo no Authorization Service.",
                "tags": [
                    "perfil"
                ],
                "summary": "Remove o próprio perfil",
                "responses": {
                    "204": {
                        "description": "Removido (ou já não existia)"
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restrito a contas com papel ADMIN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Busca um perfil por id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id do perfil (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Client"
                        }
                    },
                    "400": {
                        "description": "Id não é um UUID válido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "403": {
                        "description": "A conta não é ADMIN",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    },
                    "404": {
                        "description": "Perfil nao encontrado",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.ErroResposta"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infra"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Back-End-Cloud-Computing_client_internal_domain.Client": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "atualizadoEm": {
                    "type": "string"
                },
                "criadoEm": {
                    "type": "string"
                },
                "endereco": {
                    "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Endereco"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            }
        },
        "github_com_Back-End-Cloud-Computing_client_internal_domain.DadosPerfil": {
            "type": "object",
            "properties": {
                "endereco": {
                    "$ref": "#/definitions/github_com_Back-End-Cloud-Computing_client_internal_domain.Endereco"
                },
                "nome": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            }
        },
        "github_com_Back-End-Cloud-Computing_client_internal_domain.Endereco": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "cidade": {
                    "type": "string"
                },
                "complemento": {
                    "type": "string"
                },
                "estado": {
                    "type": "string"
                },
                "logradouro": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.ErroResposta": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token JWT emitido pelo Authorization Service.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}